created game.js a broke out necessary lines for refactor
