+++ lib/zone.js
@@ -11,7 +11,7 @@ Zone.prototype = {
     ctx.fillRect(this.x, this.y, this.height, this.width);
     return this;
   },
-  centerLine: function(ctx) {
+  drawCenterLine: function(ctx) {
     ctx.fillStyle = '#CC9900';
     ctx.fillRect(this.x, this.y, this.height, this.width);
     return this;
